+++ app/screens/competition_game.py
@@ -13,7 +13,7 @@ from typing import cast
 from helper.file_handling import (
     load_total_competition_questions,
     save_to_scores_file,
-    get_scores_key,
+    get_score_value,
 )
 from helper.settings import Settings
 from helper.game_class import GameCore, CompetitionQuestion
@@ -127,7 +127,7 @@ class Bewerb_Game(Screen):
         #     answers=["dummy answers"],
         # )
 
-        self.current_high_score = get_scores_key(
+        self.current_high_score = get_score_value(
             firetruck=self.selected_competition,
             key="high_score",
             questions="competitions",