+++ app/screens/firetruck_training.py
@@ -1,7 +1,5 @@
 from kivy.uix.button import Button
 from kivy.uix.label import Label
-from kivy.uix.image import Image
-from kivy.uix.floatlayout import FloatLayout
 from kivy.uix.screenmanager import Screen
 from kivy.clock import Clock
 
@@ -9,9 +7,10 @@ from random import shuffle
 from typing import cast
 
 from helper.functions import get_ToolQuestion_instances
-from helper.file_handling import save_to_scores_file, get_scores_key
+from helper.file_handling import save_to_scores_file, get_score_value
 from helper.settings import Settings, Strings
 from helper.game_class import GameCore
+from helper.firetruck_layouts import build_answer_layout
 
 
 settings = Settings()
@@ -59,128 +58,13 @@ class Fahrzeugkunde_Training(Screen):
 
         self.reset_strike()
 
-        self.current_high_strike = get_scores_key(
+        self.current_high_strike = get_score_value(
             self.selected_firetruck, "high_strike"
         )
 
         self.update_high_strike_label()
 
         self.next_tool()
-
-    def get_7_rooms_layout(self) -> list[tuple[str, float, float, float, float]]:
-        return [
-            ("Fahrer / GK", 0.5, 0.15, 0.25, 1),
-            ("Mannschaft", 0.5, 0.16, 0.25, 0.85),
-            ("G1", 0.27, 0.175, 0.085, 0.68),
-            ("G3", 0.205, 0.175, 0.15, 0.505),
-            ("G5", 0.27, 0.175, 0.085, 0.33),
-            ("G2", 0.27, 0.175, 0.65, 0.68),
-            ("G4", 0.205, 0.175, 0.65, 0.505),
-            ("G6", 0.27, 0.175, 0.65, 0.33),
-            ("G7 / Heck", 0.4, 0.145, 0.3, 0.145),
-            ("Dach", 0.2, 0.4, 0.4, 0.61),
-        ]
-
-    def get_5_rooms_layout(self) -> list[tuple[str, float, float, float, float]]:
-        return [
-            ("Fahrer / GK", 0.5, 0.15, 0.25, 1),
-            ("Mannschaft", 0.5, 0.16, 0.25, 0.85),
-            ("G1", 0.27, 0.175, 0.085, 0.68),
-            ("G3", 0.205, 0.35, 0.15, 0.505),
-            ("G2", 0.27, 0.175, 0.65, 0.68),
-            ("G4", 0.205, 0.35, 0.65, 0.505),
-            ("G5 / Heck", 0.4, 0.145, 0.3, 0.145),
-            ("Dach", 0.2, 0.4, 0.4, 0.61),
-        ]
-
-    def get_leiter_layout(self):
-        return [
-            ("Korb", 0.3, 0.13, 0.35, 1),
-            ("Mannschaft", 0.55, 0.155, 0.225, 0.86),
-            ("G1", 0.27, 0.2, 0.085, 0.68),
-            ("G3", 0.185, 0.185, 0.085, 0.45),
-            ("G5", 0.185, 0.125, 0.085, 0.14),
-            ("G2", 0.27, 0.2, 0.65, 0.68),
-            ("G4", 0.185, 0.185, 0.735, 0.45),
-            ("G6", 0.185, 0.125, 0.735, 0.14),
-            ("Dach", 0.28, 0.43, 0.43, 0.45),
-        ]
-
-    def get_voraus_layout(self):
-        return [
-            ("Fahrer / GK", 0.55, 0.13, 0.225, 0.90),
-            ("Mannschaft", 0.55, 0.13, 0.225, 0.75),
-            ("G1", 0.22, 0.415, 0.18, 0.61),
-            ("G1\ntief", 0.1, 0.155, 0.08, 0.35),
-            ("G2", 0.22, 0.415, 0.605, 0.61),
-            ("G2\ntief", 0.1, 0.155, 0.825, 0.35),
-            ("G3", 0.4, 0.185, 0.3, 0.19),
-        ]
-
-    def get_ruest_layout(self):
-        return [
-            ("Mannschaft", 0.55, 0.13, 0.225, 0.99),
-            ("G1\nlinks", 0.15, 0.22, 0.08, 0.81),
-            ("G1\ninnen", 0.15, 0.45, 0.23, 0.81),
-            ("G1\nrechts", 0.15, 0.22, 0.08, 0.58),
-            ("G3", 0.23, 0.265, 0.15, 0.355),
-            ("G2\nrechts", 0.15, 0.22, 0.77, 0.81),
-            ("G2\ninnen", 0.15, 0.45, 0.62, 0.81),
-            ("G2\nlinks", 0.15, 0.22, 0.77, 0.58),
-            ("G4", 0.23, 0.265, 0.62, 0.355),
-            ("G5 / Heck", 0.4, 0.08, 0.3, 0.085),
-            ("Dach", 0.2, 0.4, 0.4, 0.61),
-        ]
-
-    def build_answer_layout(self, firetruck: str) -> FloatLayout:
-        # display background and buttons
-        float = FloatLayout(size_hint=(1, 1))
-
-        if firetruck in ["RüstLösch", "Tank1", "TankDürrnberg"]:
-            bgd_image = "./assets/layouts/truck.jpg"
-            buttons = self.get_7_rooms_layout()
-
-        elif firetruck in ["Pumpe", "PumpeDürrnberg"]:
-            bgd_image = "./assets/layouts/truck.jpg"
-            buttons = self.get_5_rooms_layout()
-
-        elif firetruck in ["Leiter"]:
-            bgd_image = "./assets/layouts/leiter.jpg"
-            buttons = self.get_leiter_layout()
-
-        elif firetruck in ["Rüst"]:
-            bgd_image = "./assets/layouts/ruest.jpg"
-            buttons = self.get_ruest_layout()
-
-        elif firetruck in ["Voraus"]:
-            bgd_image = "./assets/layouts/voraus.jpg"
-            buttons = self.get_voraus_layout()
-
-        else:
-            raise NotImplementedError
-
-        background = Image(
-            source=bgd_image,
-            fit_mode="fill",
-            size_hint=(1, 1),
-            pos_hint={"center_x": 0.5, "center_y": 0.5},
-        )
-        float.add_widget(background)
-
-        # room setup for RüstLösch == TestTruck
-
-        for text, w, h, x, top in buttons:
-            btn = Button(
-                text=text,
-                font_size="25sp",
-                background_color=(0.7, 0.7, 0.7, 0.7),
-                size_hint=(w, h),
-                pos_hint={"x": x, "top": top},
-            )
-            btn.bind(on_press=self.on_answer)
-            float.add_widget(btn)
-
-        return float
 
     def next_tool(self, *args):
         self.accept_answers = True  # Enable answer processing for the new tool
@@ -199,12 +83,7 @@ class Fahrzeugkunde_Training(Screen):
 
         self.tool_label.text = self.current_tool_question.tool
 
-        # for storage in self.firetruck_rooms:
-        #     btn = Button(text=storage, font_size="28sp", disabled=storage == "")
-        #     btn.bind(on_press=self.on_answer)
-        #     self.ids.firetruck_rooms_layout.add_widget(btn)
-
-        float = self.build_answer_layout(self.selected_firetruck)
+        float = build_answer_layout(self.selected_firetruck, "fahrzeugkunde_training")
 
         self.ids.firetruck_rooms_layout.add_widget(float)
 
@@ -245,8 +124,6 @@ class Fahrzeugkunde_Training(Screen):
         else:
             self.incorrect_answer()
 
-        # children = self.firetruck_rooms_layout.children
-        # children = self.ids.firetruck_rooms_layout.children  # children is reversed
         float_layout = self.ids.firetruck_rooms_layout.children[
             0
         ]  # children is reversed
@@ -282,13 +159,6 @@ class Fahrzeugkunde_Training(Screen):
             else:
                 # answer in correct answers
                 instance.background_color = (0, 0, 1, 1)
-
-                # display hint for multiple answers
-                # if self.tool_label.text[-7:] == strings.HINT_STR_MULTIPLE_ANSWERS:
-                #     self.tool_label.text += " "
-                # else:
-                #     self.tool_label.text += "\n"
-                # self.tool_label.text += strings.HINT_STR_MULTIPLE_ANSWERS
 
                 self.tool_label.text += "\n"
                 self.tool_label.text += strings.HINT_STR_MULTIPLE_ANSWERS