+++ app/screens/firetruck_training_new.py
@@ -10,7 +10,7 @@ from random import shuffle
 from typing import cast
 
 from helper.functions import get_ToolQuestion_instances
-from helper.file_handling import save_to_scores_file, get_scores_key
+from helper.file_handling import save_to_scores_file, get_score_value
 from helper.settings import Settings, Strings
 from helper.game_class import GameCore
 
@@ -60,7 +60,7 @@ class Fahrzeugkunde_Training_New(Screen):
 
         self.reset_strike()
 
-        self.current_high_strike = get_scores_key(
+        self.current_high_strike = get_score_value(
             self.selected_firetruck, "high_strike"
         )